% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjsonwidget.R
\name{hexjsonbasefiles}
\alias{hexjsonbasefiles}
\title{List base hexjson files}
\usage{
hexjsonbasefiles()
}
\description{
\code{hexjsonbasefiles} lists the hexjson files available as part of the package.
}
\details{
The files are stored in the \code{extdata/} directory in the package (\code{inst/extadata} in the original source code).

To use one of the base files, pass the filename as the \code{jsonbase} parameter value when calling \code{hexjsonwidget}.
}
