% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjsonwidget.R
\name{hexjsondatamerge}
\alias{hexjsondatamerge}
\title{Annotate a hexjson file}
\usage{
hexjsondatamerge(jsondata, customdata, keyid = "id")
}
\arguments{
\item{jsondata}{hexJSON object}

\item{customdata}{Dataframe containing data to merge with hexjson file}

\item{keyid}{Name of \code{customdata} column for ID info}
}
\value{
a hexjson (JSON) object
}
\description{
Annotate a hexjson file
}
\author{
Tony Hirst (@psychemedia)
}
