% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjsonwidget.R
\name{hexjsonfromdataframe}
\alias{hexjsonfromdataframe}
\title{Create a hexjson file from a dataframe}
\usage{
hexjsonfromdataframe(df, layout = "odd-r", keyid = "id", q = "q",
  r = "r")
}
\arguments{
\item{df}{Dataframe to convert to hexjson}

\item{layout}{Can be one of \code{odd-r} (pointy-topped, default), \code{even-r} (pointy-topped), \code{odd-q} (flat-topped), \code{even-q} (flat-topped).}

\item{keyid}{The coloumn specifying the hex identifier/key (default is \code{id}).}

\item{q}{The column specifying the hexJSON columns (default is \code{q}).}

\item{r}{The column specifying the hexJSON rows (default is \code{r}).}
}
\value{
a hexjson (JSON) object
}
\description{
Create a hexjson file from a dataframe
}
\author{
Tony Hirst (@psychemedia)
}
