% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjsonwidget.R
\name{hexjsonread}
\alias{hexjsonread}
\title{Read hexjson file}
\usage{
hexjsonread(jsonpath)
}
\arguments{
\item{jsonpath}{path to a hexjson file}
}
\value{
a hexjson object
}
\description{
Read hexjson file
}
\author{
Tony Hirst (@psychemedia)
}
