% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjsonwidget.R
\name{hexjsonwidget}
\alias{hexjsonwidget}
\title{An htmlwidget for rendering hexjson maps using d3-hexjson}
\usage{
hexjsonwidget(jsondata = NA, jsonpath = NA, jsonbase = NA, data = NA,
  dataid = "id", datacolour = "col", datalabel = "label", grid = "off",
  labels = "on", missinglabel = NA, col_gridfill = "", col_hexfill = "",
  col_textfill = "", width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{jsondata}{hexJSON object}

\item{jsonpath}{Path to hexJSON file}

\item{jsonbase}{Name of bundled hexJSON file}

\item{data}{Dataframe containing data to merge with hexjson file}

\item{dataid}{Name of data column for ID info}

\item{datacolour}{Name of data column for colour info}

\item{datalabel}{Name of data column for label info}

\item{grid}{Display a background grid in whitespace (either \code{on} or \code{off} (default)).}

\item{labels}{Enable/Disable lables (either \code{on} (default) or \code{off})}

\item{col_gridfill}{Fill colour for grid hex}

\item{col_hexfill}{Fill colour for data hex if "col" hex value not set}

\item{col_textfill}{Fill colour for hex text label}

\item{width}{Widget width}

\item{height}{Widget height}

\item{elementId}{Widget element ID}
}
\description{
\code{hexjsonwidget()} accepts a hexJSON object or path to a hexjson file and render it as an HTMLwidget using d3-hexjson.
}
\details{
If a hexJSON hexobject includes a \code{col} attribute, this will be used to colour the hex.
If it contains a \code{label} attribute, it will be used to label the hex.
}
\examples{
hexjsonwidget( fromJSON('./example.hexjson') )
hexjsonwidget(jsonpath='./example-grid.hexjson')
hexjsonwidget(jj, grid='on', col_gridfill='#113355', col_hexfill='#FF0000')
hexjsonwidget(jj, grid='off', col_textfill="orange")
hexjsonwidget(jj, col_hexfill='#bb3388')

}
\author{
Tony Hirst (@psychemedia)
}
