% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjsonwidget.R
\name{hexjsonwrite}
\alias{hexjsonwrite}
\title{Write hexjson file from a hexjson object}
\usage{
hexjsonwrite(jsondata, outfilename)
}
\arguments{
\item{jsondata}{hexJSON object}
}
\description{
Write hexjson file from a hexjson object
}
\author{
Tony Hirst (@psychemedia)
}
