% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjsonwidget.R
\name{hexjsondatamerge}
\alias{hexjsondatamerge}
\title{Annotate a hexjson file}
\usage{
hexjsondatamerge(jsondata, customdata, dataid = "id", datacolour = "col",
  datalabel = "label")
}
\arguments{
\item{jsondata}{hexJSON object}

\item{customdata}{Dataframe containing data to merge with hexjson file}

\item{dataid}{Name of \code{customdata} column for ID info}

\item{datacolour}{Name of \code{customdata} column for colour info}

\item{datalabel}{Name of \code{customdata} column for label info}
}
\description{
Annotate a hexjson file
}
\author{
Tony Hirst (@psychemedia)
}
