% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjsonwidget.R
\docType{package}
\name{hexjsonwidget}
\alias{hexjsonwidget}
\alias{hexjsonwidget-package}
\title{An htmlwidget for rendering hexjson maps using d3-hexjson}
\usage{
hexjsonwidget(jsondata = NA, jsonpath = "", grid = "off",
  col_hexfill = "", col_gridfill = "", col_textfill = "", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{jsondata}{hexJSON object}

\item{jsonpath}{Path to hexJSON file}

\item{grid}{Display a background grid in whitespace (either \code{on} or \code{off} (default)).}

\item{col_hexfill}{Fill colour for data hex if "col" hex value not set}

\item{col_gridfill}{Fill colour for grid hex}

\item{col_textfill}{Fill colour for hex text label}

\item{width}{}

\item{elementId}{}
}
\description{
\code{hexjsonwidget} accepts a hexJSON object or path to a hexjson file and render it as an HTMLwidget using d3-hexjson.
}
\author{
Tony Hirst (@psychemedia)
}
